% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrerfHelperFunctions.R
\name{BICCutMclust}
\alias{BICCutMclust}
\title{Find minimizing BIC Cut for Vector}
\usage{
BICCutMclust(X)
}
\arguments{
\item{X}{a one dimensional vector}
}
\value{
list containing minimizing cut point and corresponding BIC score.
}
\description{
Find minimizing BIC Cut for Vector
}
