% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getFeatures}
\alias{getFeatures}
\title{Extract feature indicies from the sparse projection vector.}
\usage{
getFeatures(x)
}
\arguments{
\item{a}{list of unique.projections from the intermediate steps of
the FeatureImportance function.}
}
\value{
list of unique feature combinations
}
\description{
A helper function to extract the feature indices from the projection
vector stored in a tree object.
}
