% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrerfHelperFunctions.R
\name{makeAB}
\alias{makeAB}
\title{Create rotation matrix used to determine mtry features.}
\usage{
makeAB(options)
}
\arguments{
\item{options}{a list of options used by the sparse matrix generator.  These can vary if a different matrix is required.}
}
\value{
rotationMatrix the matrix used to determine which mtry features or combination of features will be used to split a node.
}
\description{
Create rotation matrix used to determine mtry features.
}
