% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackForest.R
\name{PackForest}
\alias{PackForest}
\title{Compute class predictions for each observation in X}
\usage{
PackForest(X, Y, forest)
}
\arguments{
\item{X}{an n by d numeric matrix (preferable) or data frame. The rows correspond to observations and columns correspond to features of a test set, which should be different from the training set.}

\item{Y}{the number of cores to use while training. If NumCores=0 then 1 less than the number of cores reported by the OS are used. (NumCores=0)}

\item{forest}{a forest trained using the RerF function.}
}
\value{
predictions an n length vector of predictions
}
\description{
Predicts the classification of samples using a trained forest.
}
\examples{
library(rerf)
trainIdx <- c(1:40, 51:90, 101:140)
X <- as.matrix(iris[,1:4])
Y <- as.numeric(iris[,5])

}
