% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandMatCat.R
\name{RandMatCat}
\alias{RandMatCat}
\title{Create a Random Matrix for use when categorical features are present}
\usage{
RandMatCat(mat.options)
}
\arguments{
\item{mat.options}{the same as that for RandMat, except an additional fifth argument cat.map is taken, which specifies which one-of-K encoded columns in X correspond to the same categorical feature.}
}
\value{
random.matrix
}
\description{
The same as the function RandMat, except that this function is used when a cat.map object is provided (see RerF and GetCatMap for details).
}
\author{
James Browne (jbrowne6@jhu.edu) and Tyler Tomita (ttomita2@jhmi.edu)
}
