% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandMat.R
\name{RandMatTSpatch}
\alias{RandMatTSpatch}
\title{Create a Random Matrix: ts-patch}
\usage{
RandMatTSpatch(p, d, pwMin, pwMax, ...)
}
\arguments{
\item{p}{the number of dimensions.}

\item{d}{the number of desired columns in the projection matrix.}

\item{pwMin}{the minimum patch size to sample.}

\item{pwMax}{the maximum patch size to sample.}

\item{...}{used to handle superfluous arguments passed in using paramList.}
}
\value{
A random matrix to use in running \code{\link{RerF}}.
}
\description{
Create a Random Matrix: ts-patch
}
\examples{

p <- 8
d <- 8
pwMin <- 3
pwMax <- 6
paramList <- list(p = p, d = d, pwMin = pwMin, pwMax = pwMax)
set.seed(8)
(a <- do.call(RandMatTSpatch, paramList))

}
