% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFeatureImportance.R
\name{RunFeatureImportance}
\alias{RunFeatureImportance}
\title{Compute Feature Importance of a single RerF tree}
\usage{
RunFeatureImportance(tree, unique.projections)
}
\arguments{
\item{tree}{a single tree from a trained RerF model with argument store.impurity = TRUE.}

\item{unique.projections}{a list of all of the unique split projections used in the RerF model.}
}
\value{
feature.imp
}
\description{
Computes feature importance of every unique feature used to make a split in a single tree.
}
