% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunOOB.R
\name{RunOOB}
\alias{RunOOB}
\title{Predict class labels on out-of-bag observations using a single tree.}
\usage{
RunOOB(X, tree)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame which was used to train the provided forest.}

\item{tree}{a tree from a forest returned by RerF.}
}
\value{
predictions prediction matrix used by OOBPredict
}
\description{
This is the base function called by OOBPredict.
}
