% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleNode.R
\name{Scale01}
\alias{Scale01}
\title{Scale data in current Node to [0,1]}
\usage{
Scale01(Xnode, scalingFactors = NULL)
}
\arguments{
\item{Xnode}{an n by d numeric matrix or data frame containing the data assigned to the current node. The rows correspond to observations and columns correspond to features.}

\item{scalingFactors}{a list with names elements \code{min} and
\code{diff} containing vectors of the minima and difference in maxima
and minima across columns of Xnode.}
}
\value{
list with the scaled data and scaling factors
}
\description{
Scales the data in the current node to live in [0,1] over the columns.
}
\examples{
set.seed(12345) 
X <- as.matrix(iris[, -5])
CurrentNode <- 1 
Assigned2Node <- list(sample(1:nrow(X), 10))

sparseM <- RandMatBinary(p = 4, d = 2, sparsity = 0.5, prob = 0.5)

lrows  <- which(sparseM[, 2L] == 1)
Xnode <- X[Assigned2Node[[CurrentNode]], sparseM[lrows, 2L, drop = FALSE]]
scaleDat <- rerf:::Scale01(X[Assigned2Node[[CurrentNode]], , drop = FALSE])

}
