% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{uniqueByEquivalenceClass}
\alias{uniqueByEquivalenceClass}
\title{Remove Unique Projections that are equivalent due to a rotation of 180
degrees.}
\usage{
uniqueByEquivalenceClass(unique.projections)
}
\arguments{
\item{unique.projection}{a list of projections from intermediate
steps of the \code{\link{FeatureImportance}} function.}
}
\value{
a list which is a subset of the input.
}
\description{
This function finds the projections that are equivalent via a 180
degree rotation and removes the duplicates.
}
\seealso{
\code{\link{FeatureImportance}}
}
