% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnsupervisedLumberjack.R
\name{UnsupervisedLumberjack}
\alias{UnsupervisedLumberjack}
\title{Unsupervised Lumberjack}
\usage{
UnsupervisedLumberjack(X, trees = 100, min.parent = 10,
  max.depth = NA)
}
\arguments{
\item{X}{an n by d numeric matrix (preferable) or data frame. The rows correspond to observations and columns correspond to features.  X must be numeric or convertable to numeric.}

\item{trees}{the number of trees to grow (trees = 100).}

\item{min.parent}{the minimum leaf node size.  Use of min.parent
increases liklihood that nodes will contain multiple observations.
e.g. (min.parent = 10)}

\item{max.depth}{the longest allowable distance from the root of a
tree to a leaf node (i.e. the maximum allowed height for a tree).  If
depth=NA then min.parent is used as the sole stopping criteria. (depth=NA)}
}
\value{
urerfStructure
}
\description{
Creates the forest and supporting data that makes up the uLumberjack structure. At its core this function is a wrapper for the GrowUnsupervisedForest function.
}
\examples{
library(rerf)
X <- as.matrix(iris[,1:4])
urerfStructure <- UnsupervisedLumberjack(X, trees=100,
                                         min.parent=10, max.depth=5)

}
