% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandMat.R
\name{RandMatBinary}
\alias{RandMatBinary}
\title{Create a Random Matrix: Binary}
\usage{
RandMatBinary(p, d, rho, prob, catMap = NULL)
}
\arguments{
\item{p}{the number of dimensions.}

\item{d}{the number of desired columns in the projection matrix.}

\item{rho}{a real number in \eqn{(0,1)} that specifies the distibution of
non-zero elements in the random matrix.}

\item{prob}{a probability \eqn{\in (0,1)} used for creating the random
matrix.}

\item{catMap}{a list specifying specifies which one-of-K encoded columns in X correspond to the same categorical feature.}
}
\value{
A random matrix to use in running \code{\link{RerF}}.
}
\description{
Create a Random Matrix: Binary
}
\examples{

p <- 8
d <- 3
rho <- 0.25
prob <- 0.5
set.seed(4)
(a <- RandMatBinary(p, d, rho, prob))

}
