% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{uniqueByEquivalenceClass}
\alias{uniqueByEquivalenceClass}
\title{Remove unique projections that are equivalent due to a rotation of 180
degrees.}
\usage{
uniqueByEquivalenceClass(p, unique.projections)
}
\arguments{
\item{p}{the number of features in the original data.  This can be
obtained from a forest object via \code{forest$params$paramList$p}.}

\item{unique.projections}{a list of projections from intermediate
steps of the \code{\link{FeatureImportance}} function.}
}
\value{
unique.projections a list which is a subset of the input.
}
\description{
This function finds the projections that are equivalent via a 180
degree rotation and removes the duplicates.
}
\seealso{
\code{\link{FeatureImportance}}
}
